Add new glass color-effect. Abstract navbar component.
